pallet_parachain_staking
Module inflation

Helper methods for computing issuance based on inflation

Structs

InflationInfo

Range

Functions

annual_to_round

Convert annual inflation rate range to round inflation range

perbill_annual_to_perbill_round

Convert an annual inflation to a round inflation round = (1+annual)^(1/rounds_per_year) - 1

round_issuance_range

Compute round issuance range from round inflation range and current total issuance