pallet_parachain_staking
Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

AddGet

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

Pallet for parachain staking

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

FreezeReason

The reason for freezing funds.

Constants

COLLATOR_LOCK_ID

DELEGATOR_LOCK_ID

MAX_CANDIDATES

A hard limit for weight computation purposes for the max candidates that could theoretically exist.

Traits

Config

Configuration trait of this pallet.

Type Aliases

AtStake

Snapshot of collator delegation stake at the start of the round

AwardedPts

Points for each collator per round

BalanceOf

DelayedPayouts

Delayed payouts

EnableMarkingOffline

Killswitch to enable/disable marking offline feature.

InflationConfig

Inflation configuration

MigratedCandidates

Temporary storage to track candidates that have been migrated from locks to freezes. This storage should be removed after all accounts have been migrated.

MigratedDelegators

Temporary storage to track delegators that have been migrated from locks to freezes. This storage should be removed after all accounts have been migrated.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Points

Total points awarded to collators for block production in the round

Round

Current round index and next round scheduled transition

RoundIndex

WasInactive

Records collators’ inactivity. Data persists for MaxOfflineRounds + 1 rounds before being pruned.