pallet_parachain_staking::pallet::dispatchables
Function migrate_locks_to_freezes_batch

pub fn migrate_locks_to_freezes_batch<T: Config>(
    accounts: BoundedVec<(T::AccountId, bool), ConstU32<MAX_ACCOUNTS_PER_MIGRATION_BATCH>>,
)

Batch migrate locks to freezes for a list of accounts.

This function allows migrating multiple accounts from the old lock-based staking to the new freeze-based staking in a single transaction.

Parameters:

accounts: List of tuples containing (account_id, is_collator) where is_collator indicates if the account is a collator (true) or delegator (false)

The maximum number of accounts that can be migrated in one batch is MAX_ACCOUNTS_PER_MIGRATION_BATCH. The batch cannot be empty.

If 50% or more of the migration attempts are successful, the entire extrinsic fee is refunded to incentivize successful batch migrations. Weight is calculated based on actual successful operations performed.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::migrate_locks_to_freezes_batch.