pallet_parachain_staking::pallet
Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AtStake

Snapshot of collator delegation stake at the start of the round

AutoCompoundingDelegations

Stores auto-compounding configuration per collator.

AwardedPts

Points for each collator per round

BottomDelegations

Bottom delegations for collator candidate

CandidateInfo

Get collator candidate info associated with an account if account is candidate else None

CandidatePool

The pool of collator candidates, each with their total backing stake

CollatorCommission

Commission percent taken off of rewards for all collators

DelayedPayouts

Delayed payouts

DelegationScheduledRequests

Stores outstanding delegation requests per collator.

DelegatorState

Get delegator state associated with an account if account is delegating else None

EnableMarkingOffline

Killswitch to enable/disable marking offline feature.

InflationConfig

Inflation configuration

InflationDistributionInfo

Inflation distribution configuration, including accounts that should receive inflation before it is distributed to collators and delegators.

Points

Total points awarded to collators for block production in the round

Round

Current round index and next round scheduled transition

SelectedCandidates

The collator candidates selected for the current round

TopDelegations

Top delegations for collator candidate

Total

Total capital locked by this staking pallet

TotalSelected

The total candidates selected every round

WasInactive

Records collators’ inactivity. Data persists for MaxOfflineRounds + 1 rounds before being pruned.