Struct pallet_parachain_staking::pallet::Pallet

pub struct Pallet<T>(/* private fields */);

Pallet for parachain staking

Implementations

impl<T: Config> Pallet<T>

pub fn delegation_request_exists( collator: &T::AccountId, delegator: &T::AccountId, ) -> bool

Returns true if a ScheduledRequest exists for a given delegation

pub fn delegation_request_revoke_exists( collator: &T::AccountId, delegator: &T::AccountId, ) -> bool

Returns true if a DelegationAction::Revoke ScheduledRequest exists for a given delegation

impl<T: Config> Pallet<T>

pub fn set_staking_expectations( origin: OriginFor<T>, expectations: Range<BalanceOf<T>>, ) -> DispatchResultWithPostInfo

Set the expectations for total staked. These expectations determine the issuance for the round according to logic in fn compute_issuance

pub fn set_inflation( origin: OriginFor<T>, schedule: Range<Perbill>, ) -> DispatchResultWithPostInfo

Set the annual inflation rate to derive per-round inflation

pub fn set_parachain_bond_account( origin: OriginFor<T>, new: T::AccountId, ) -> DispatchResultWithPostInfo

Deprecated: please use set_inflation_distribution_config instead.

Set the account that will hold funds set aside for parachain bond

pub fn set_parachain_bond_reserve_percent( origin: OriginFor<T>, new: Percent, ) -> DispatchResultWithPostInfo

Deprecated: please use set_inflation_distribution_config instead.

Set the percent of inflation set aside for parachain bond

pub fn set_total_selected( origin: OriginFor<T>, new: u32, ) -> DispatchResultWithPostInfo

Set the total number of collator candidates selected per round

changes are not applied until the start of the next round

pub fn set_collator_commission( origin: OriginFor<T>, new: Perbill, ) -> DispatchResultWithPostInfo

Set the commission for all collators

pub fn set_blocks_per_round( origin: OriginFor<T>, new: u32, ) -> DispatchResultWithPostInfo

Set blocks per round

if called with new less than length of current round, will transition immediately in the next block

also updates per-round inflation config

pub fn join_candidates( origin: OriginFor<T>, bond: BalanceOf<T>, candidate_count: u32, ) -> DispatchResultWithPostInfo

Join the set of collator candidates

pub fn schedule_leave_candidates( origin: OriginFor<T>, candidate_count: u32, ) -> DispatchResultWithPostInfo

Request to leave the set of candidates. If successful, the account is immediately removed from the candidate pool to prevent selection as a collator.

pub fn execute_leave_candidates( origin: OriginFor<T>, candidate: T::AccountId, candidate_delegation_count: u32, ) -> DispatchResultWithPostInfo

Execute leave candidates request

pub fn cancel_leave_candidates( origin: OriginFor<T>, candidate_count: u32, ) -> DispatchResultWithPostInfo

Cancel open request to leave candidates

only callable by collator account

result upon successful call is the candidate is active in the candidate pool

pub fn go_offline(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Temporarily leave the set of collator candidates without unbonding

pub fn go_online(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Rejoin the set of collator candidates if previously had called go_offline

pub fn candidate_bond_more( origin: OriginFor<T>, more: BalanceOf<T>, ) -> DispatchResultWithPostInfo

Increase collator candidate self bond by more

pub fn schedule_candidate_bond_less( origin: OriginFor<T>, less: BalanceOf<T>, ) -> DispatchResultWithPostInfo

Request by collator candidate to decrease self bond by less

pub fn execute_candidate_bond_less( origin: OriginFor<T>, candidate: T::AccountId, ) -> DispatchResultWithPostInfo

Execute pending request to adjust the collator candidate self bond

pub fn cancel_candidate_bond_less( origin: OriginFor<T>, ) -> DispatchResultWithPostInfo

Cancel pending request to adjust the collator candidate self bond

pub fn delegate_with_auto_compound( origin: OriginFor<T>, candidate: T::AccountId, amount: BalanceOf<T>, auto_compound: Percent, candidate_delegation_count: u32, candidate_auto_compounding_delegation_count: u32, delegation_count: u32, ) -> DispatchResultWithPostInfo

If caller is not a delegator and not a collator, then join the set of delegators If caller is a delegator, then makes delegation to change their delegation state Sets the auto-compound config for the delegation

pub fn schedule_revoke_delegation( origin: OriginFor<T>, collator: T::AccountId, ) -> DispatchResultWithPostInfo

Request to revoke an existing delegation. If successful, the delegation is scheduled to be allowed to be revoked via the execute_delegation_request extrinsic. The delegation receives no rewards for the rounds while a revoke is pending. A revoke may not be performed if any other scheduled request is pending.

pub fn delegator_bond_more( origin: OriginFor<T>, candidate: T::AccountId, more: BalanceOf<T>, ) -> DispatchResultWithPostInfo

Bond more for delegators wrt a specific collator candidate.

pub fn schedule_delegator_bond_less( origin: OriginFor<T>, candidate: T::AccountId, less: BalanceOf<T>, ) -> DispatchResultWithPostInfo

Request bond less for delegators wrt a specific collator candidate. The delegation’s rewards for rounds while the request is pending use the reduced bonded amount. A bond less may not be performed if any other scheduled request is pending.

pub fn execute_delegation_request( origin: OriginFor<T>, delegator: T::AccountId, candidate: T::AccountId, ) -> DispatchResultWithPostInfo

Execute pending request to change an existing delegation

pub fn cancel_delegation_request( origin: OriginFor<T>, candidate: T::AccountId, ) -> DispatchResultWithPostInfo

Cancel request to change an existing delegation.

pub fn set_auto_compound( origin: OriginFor<T>, candidate: T::AccountId, value: Percent, candidate_auto_compounding_delegation_count_hint: u32, delegation_count_hint: u32, ) -> DispatchResultWithPostInfo

Sets the auto-compounding reward percentage for a delegation.

pub fn hotfix_remove_delegation_requests_exited_candidates( origin: OriginFor<T>, candidates: Vec<T::AccountId>, ) -> DispatchResult

Hotfix to remove existing empty entries for candidates that have left.

pub fn notify_inactive_collator( origin: OriginFor<T>, collator: T::AccountId, ) -> DispatchResult

Notify a collator is inactive during MaxOfflineRounds

pub fn enable_marking_offline( origin: OriginFor<T>, value: bool, ) -> DispatchResult

Enable/Disable marking offline feature

pub fn force_join_candidates( origin: OriginFor<T>, account: T::AccountId, bond: BalanceOf<T>, candidate_count: u32, ) -> DispatchResultWithPostInfo

Force join the set of collator candidates. It will skip the minimum required bond check.

pub fn set_inflation_distribution_config( origin: OriginFor<T>, new: InflationDistributionConfig<T::AccountId>, ) -> DispatchResultWithPostInfo

Set the inflation distribution configuration.

impl<T: Config> Pallet<T>

pub fn set_candidate_bond_to_zero(acc: &T::AccountId) -> Weight

pub fn is_delegator(acc: &T::AccountId) -> bool

pub fn is_candidate(acc: &T::AccountId) -> bool

pub fn is_selected_candidate(acc: &T::AccountId) -> bool

pub fn join_candidates_inner( acc: T::AccountId, bond: BalanceOf<T>, candidate_count: u32, ) -> DispatchResultWithPostInfo

pub fn go_offline_inner(collator: T::AccountId) -> DispatchResultWithPostInfo

pub fn go_online_inner(collator: T::AccountId) -> DispatchResultWithPostInfo

pub fn candidate_bond_more_inner( collator: T::AccountId, more: BalanceOf<T>, ) -> DispatchResultWithPostInfo

pub fn execute_candidate_bond_less_inner( candidate: T::AccountId, ) -> DispatchResultWithPostInfo

pub fn execute_leave_candidates_inner( candidate: T::AccountId, ) -> DispatchResultWithPostInfo

pub fn get_delegator_stakable_balance(acc: &T::AccountId) -> BalanceOf<T>

Returns an account’s stakable balance which is not locked in delegation staking

pub fn get_collator_stakable_free_balance(acc: &T::AccountId) -> BalanceOf<T>

Returns an account’s free balance which is not locked in collator staking

pub fn delegation_auto_compound( candidate: &T::AccountId, delegator: &T::AccountId, ) -> Percent

Returns a delegations auto-compound value.

pub fn compute_top_candidates() -> Vec<T::AccountId>

Compute the top TotalSelected candidates in the CandidatePool and return a vec of their AccountIds (sorted by AccountId).

If the returned vec is empty, the previous candidates should be used.

pub fn delegation_bond_more_without_event( delegator: T::AccountId, candidate: T::AccountId, more: BalanceOf<T>, ) -> Result<(bool, Weight), DispatchErrorWithPostInfo<PostDispatchInfo>>

This function exists as a helper to delegator_bond_more & auto_compound functionality. Any changes to this function must align with both user-initiated bond increases and auto-compounding bond increases. Any feature-specific preconditions should be validated before this function is invoked. Any feature-specific events must be emitted after this function is invoked.

pub fn mint(amt: BalanceOf<T>, to: T::AccountId)

Mint a specified reward amount to the beneficiary account. Emits the [Rewarded] event.

pub fn mint_collator_reward( _paid_for_round: RoundIndex, collator_id: T::AccountId, amt: BalanceOf<T>, ) -> Weight

Mint a specified reward amount to the collator’s account. Emits the [Rewarded] event.

pub fn mint_and_compound( amt: BalanceOf<T>, compound_percent: Percent, candidate: T::AccountId, delegator: T::AccountId, )

Mint and compound delegation rewards. The function mints the amount towards the delegator and tries to compound a specified percent of it back towards the delegation. If a scheduled delegation revoke exists, then the amount is only minted, and nothing is compounded. Emits the [Compounded] event.

pub fn mark_collators_as_inactive(cur: RoundIndex) -> Weight

Marks collators as inactive for the previous round if they received zero awarded points.

impl<T: Config> Pallet<T>

pub fn collator_commission() -> Perbill

An auto-generated getter for CollatorCommission.

impl<T: Config> Pallet<T>

pub fn total_selected() -> u32

An auto-generated getter for TotalSelected.

impl<T: Config> Pallet<T>

pub fn inflation_distribution_info() -> InflationDistributionConfig<T::AccountId>

An auto-generated getter for InflationDistributionInfo.

impl<T: Config> Pallet<T>

pub fn round() -> RoundInfo<BlockNumberFor<T>>

An auto-generated getter for Round.

impl<T: Config> Pallet<T>

pub fn delegator_state<KArg>( k: KArg, ) -> Option<Delegator<T::AccountId, BalanceOf<T>>>
where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for DelegatorState.

impl<T: Config> Pallet<T>

pub fn candidate_info<KArg>(k: KArg) -> Option<CandidateMetadata<BalanceOf<T>>>
where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for CandidateInfo.

impl<T: Config> Pallet<T>

pub fn delegation_scheduled_requests<KArg>( k: KArg, ) -> BoundedVec<ScheduledRequest<T::AccountId, BalanceOf<T>>, AddGet<T::MaxTopDelegationsPerCandidate, T::MaxBottomDelegationsPerCandidate>>
where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for DelegationScheduledRequests.

impl<T: Config> Pallet<T>

pub fn auto_compounding_delegations<KArg>( k: KArg, ) -> BoundedVec<AutoCompoundConfig<T::AccountId>, AddGet<T::MaxTopDelegationsPerCandidate, T::MaxBottomDelegationsPerCandidate>>
where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for AutoCompoundingDelegations.

impl<T: Config> Pallet<T>

pub fn top_delegations<KArg>( k: KArg, ) -> Option<Delegations<T::AccountId, BalanceOf<T>>>
where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for TopDelegations.

impl<T: Config> Pallet<T>

pub fn bottom_delegations<KArg>( k: KArg, ) -> Option<Delegations<T::AccountId, BalanceOf<T>>>
where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for BottomDelegations.

impl<T: Config> Pallet<T>

pub fn selected_candidates() -> BoundedVec<T::AccountId, T::MaxCandidates>

An auto-generated getter for SelectedCandidates.

impl<T: Config> Pallet<T>

pub fn total() -> BalanceOf<T>

An auto-generated getter for Total.

impl<T: Config> Pallet<T>

pub fn candidate_pool() -> BoundedOrderedSet<Bond<T::AccountId, BalanceOf<T>>, T::MaxCandidates>

An auto-generated getter for CandidatePool.

impl<T: Config> Pallet<T>

pub fn at_stake<KArg1, KArg2>( k1: KArg1, k2: KArg2, ) -> Option<CollatorSnapshot<T::AccountId, BalanceOf<T>>>
where KArg1: EncodeLike<RoundIndex>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for AtStake.

impl<T: Config> Pallet<T>

pub fn was_inactive<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> Option<()>
where KArg1: EncodeLike<RoundIndex>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for WasInactive.

impl<T: Config> Pallet<T>

pub fn delayed_payouts<KArg>(k: KArg) -> Option<DelayedPayout<BalanceOf<T>>>
where KArg: EncodeLike<RoundIndex>,

An auto-generated getter for DelayedPayouts.

impl<T: Config> Pallet<T>

pub fn inflation_config() -> InflationInfo<BalanceOf<T>>

An auto-generated getter for InflationConfig.

impl<T: Config> Pallet<T>

pub fn points<KArg>(k: KArg) -> u32
where KArg: EncodeLike<RoundIndex>,

An auto-generated getter for Points.

impl<T: Config> Pallet<T>

pub fn awarded_pts<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> u32
where KArg1: EncodeLike<RoundIndex>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for AwardedPts.

impl<T: Config> Pallet<T>

pub fn marking_offline() -> bool

An auto-generated getter for EnableMarkingOffline.