Module pallet_parachain_staking::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

cancel_candidate_bond_less

Cancel pending request to adjust the collator candidate self bond

cancel_delegation_request

Cancel request to change an existing delegation.

cancel_leave_candidates

Cancel open request to leave candidates

candidate_bond_more

Increase collator candidate self bond by more

delegate

DEPRECATED use delegateWithAutoCompound If caller is not a delegator and not a collator, then join the set of delegators If caller is a delegator, then makes delegation to change their delegation state

delegate_with_auto_compound

If caller is not a delegator and not a collator, then join the set of delegators If caller is a delegator, then makes delegation to change their delegation state Sets the auto-compound config for the delegation

delegator_bond_more

Bond more for delegators wrt a specific collator candidate.

enable_marking_offline

Enable/Disable marking offline feature

execute_candidate_bond_less

Execute pending request to adjust the collator candidate self bond

execute_delegation_request

Execute pending request to change an existing delegation

execute_leave_candidates

Execute leave candidates request

force_join_candidates

Force join the set of collator candidates. It will skip the minimum required bond check.

go_offline

Temporarily leave the set of collator candidates without unbonding

go_online

Rejoin the set of collator candidates if previously had called go_offline

hotfix_remove_delegation_requests_exited_candidates

Hotfix to remove existing empty entries for candidates that have left.

join_candidates

Join the set of collator candidates

notify_inactive_collator

Notify a collator is inactive during MaxOfflineRounds

removed_call_19

REMOVED, was schedule_leave_delegators

removed_call_20

REMOVED, was execute_leave_delegators

removed_call_21

REMOVED, was cancel_leave_delegators

schedule_candidate_bond_less

Request by collator candidate to decrease self bond by less

schedule_delegator_bond_less

Request bond less for delegators wrt a specific collator candidate. The delegation’s rewards for rounds while the request is pending use the reduced bonded amount. A bond less may not be performed if any other scheduled request is pending.

schedule_leave_candidates

Request to leave the set of candidates. If successful, the account is immediately removed from the candidate pool to prevent selection as a collator.

schedule_revoke_delegation

Request to revoke an existing delegation. If successful, the delegation is scheduled to be allowed to be revoked via the execute_delegation_request extrinsic. The delegation receives no rewards for the rounds while a revoke is pending. A revoke may not be performed if any other scheduled request is pending.

set_auto_compound

Sets the auto-compounding reward percentage for a delegation.

set_blocks_per_round

Set blocks per round

set_collator_commission

Set the commission for all collators

set_inflation

Set the annual inflation rate to derive per-round inflation

set_parachain_bond_account

Set the account that will hold funds set aside for parachain bond

set_parachain_bond_reserve_percent

Set the percent of inflation set aside for parachain bond

set_staking_expectations

Set the expectations for total staked. These expectations determine the issuance for the round according to logic in fn compute_issuance

set_total_selected

Set the total number of collator candidates selected per round