Module pallet_parachain_staking::pallet::storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Structs

AtStake

Snapshot of collator delegation stake at the start of the round

AutoCompoundingDelegations

Stores auto-compounding configuration per collator.

AwardedPts

Points for each collator per round

BottomDelegations

Bottom delegations for collator candidate

CandidateInfo

Get collator candidate info associated with an account if account is candidate else None

CandidatePool

The pool of collator candidates, each with their total backing stake

CollatorCommission

Commission percent taken off of rewards for all collators

DelayedPayouts

Delayed payouts

DelegationScheduledRequests

Stores outstanding delegation requests per collator.

DelegatorState

Get delegator state associated with an account if account is delegating else None

InflationConfig

Inflation configuration

ParachainBondInfo

Parachain bond config info { account, percent_of_inflation }

Points

Total points awarded to collators for block production in the round

Round

Current round index and next round scheduled transition

SelectedCandidates

The collator candidates selected for the current round

Staked

Total counted stake for selected candidates in the round

TopDelegations

Top delegations for collator candidate

Total

Total capital locked by this staking pallet

TotalSelected

The total candidates selected every round