pallet_parachain_staking
Module types

Types for parachain-staking

Structs

Bond

BondWithAutoCompound

CandidateBondLessRequest

Request scheduled to change the collator candidate self-bond

CandidateMetadata

All candidate info except the top and bottom delegations

CollatorSnapshot

Snapshot of collator state at the start of the round for which they are selected

CountedDelegations

DelayedPayout

Info needed to make delayed payments to stakers after round end

Delegations

Type for top and bottom delegation storage item

Delegator

Delegator state

InflationDistributionAccount

Reserve information { account, percent_of_inflation }

InflationDistributionConfig

RoundInfo

The current round index and transition information

Enums

BondAdjust

CapacityStatus

Capacity status for top or bottom delegations

CollatorStatus

The activity status of the collator

DelegatorAdded

Convey relevant information describing if a delegator was added to the top or bottom Delegations added to the top yield a new total

DelegatorStatus